+++ src/build-info.ts
@@ -1,2 +1,2 @@
-export const BUILD_LABEL = "ThemeDeck Build 2026-02-23 19:29:55 EST";
-export const BUILD_ID = "2026-02-23-192955-EST";
+export const BUILD_LABEL = "ThemeDeck Build 2026-02-24 11:36:37 EST";
+export const BUILD_ID = "2026-02-24-113637-EST";